+++ HISTORY.rst
@@ -8,6 +8,7 @@ Release History
 
 - Added `namespace_suffix` to EventHubConfig() to support national clouds.
 - Added `device_id` attribute to EventData to support IoT Hub use cases.
+- Added message header to workaround service bug for PartitionKey support.
 
 
 0.2.0b1 (2018-04-20)